requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-storage
psycopg2
pandas
pycountry
unidecode
google-cloud-secret-manager
lxml